--- src/utils/requests/new_user/create_user.py
@@ -1,7 +1,0 @@
-from user_details import UserDetails
-
-
-class CreateUser(UserDetails):
-
-    def __init__(self, name, description, email):
-        super().__init__(name, description, email)
