--- src/utils/requests/new_user/user_details.py
@@ -1,9 +1,0 @@
-from abc import ABC
-
-
-class UserDetails(ABC):
-
-    def __init__(self, name, description, email):
-        self.name = name
-        self.description = description
-        self.email = email
